% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratExtensions.R
\name{customPercentageFeature}
\alias{customPercentageFeature}
\title{Title}
\usage{
customPercentageFeature(AGG, mitoPattern, markerGenes, genes)
}
\arguments{
\item{genes}{}
}
\value{

}
\description{
Title
}
