% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratExtensions.R
\name{filterSeurat}
\alias{filterSeurat}
\title{Title}
\usage{
filterSeurat(Object, mito.range = c(0, 100), gene_range = c(0, Inf),
  umi_range = c(0, Inf))
}
\arguments{
\item{umi_range}{}
}
\description{
Title
}
