% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratExtensions.R
\name{read10XwithMarkergenes}
\alias{read10XwithMarkergenes}
\title{Title}
\usage{
read10XwithMarkergenes(tenxPath, pMin.cells = 3, pMin.features = 20,
  markerGenes, mitoPattern = "^mt-", projectName,
  cellRangerAggregated = TRUE)
}
\arguments{
\item{cellRangerAggregated}{}
}
\description{
Title
}
